requirements.txt
facenet_pytorch
torch
numpy
mmcv
pillow
IPython
pytube